requirements.txt
Django==4.0.1
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
pillow==9.3.0
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
django-cors-headers==3.11.0

python-dotenv==0.20.0
gunicorn==20.1.0

django-redis==5.2.0